Saved the map plot as both html and png
The html is interactive but if you are still unable to run it on your end I will also upload the output_data folder that contains both formats. Just beats in mind that the directory of the html format is for my personal desktop. But if you run it on your end it will change to yours instead.